flyttet eksperimentelle tester til en egen mappe
