+++ Eksperimenteringer/test_hev_senk/.gitignore
@@ -1,0 +1,3 @@
+__pycache__/
+*.pyc
+.venv/
