--- test_function/.gitignore
@@ -1,3 +1,0 @@
-__pycache__/
-*.pyc
-.venv/
