--- test_function/.vscode/settings.json
@@ -1,7 +1,0 @@
-// Place your settings in this file to overwrite default and user settings.
-{
-	"files.eol": "\n",
-	"debug.openDebug": "neverOpen",
-	"python.linting.enabled": false,
-	"python.languageServer": "None"
-}
